Bumping assembly version to 1.1.3
Fixed bug where the progress dialog was not closed on an exception so it stayed on screen forever.

Fixed bug in showing file dialog for settings if the starting path was a drive root.

Fixed bug where removing a channel if that was the only channel would not save between sessions.