Make sure stream launcher is a singleton for the purpose of keeping track of currently watched livestreams.
Keep track of currently watching streams in stream launcher, use this to prevent notifications for streams that are currently being watched.

Fixed LivestreamModel equality comparison to work off id instead of display name.

Make slightly more use of the stream provider to determine the base url for the stream, this will all be refactored into IStreamProvider implementations when I get around to supporting youtube (and any other popular streaming services in the future).