Fix missing popups for new livestreams
Due to addrange/removerange triggering reset events without populating what was added or removed, we have to go back to using an iterator and adding/removing each stream 1 at a time.

Fixed a bug where youtube streams would be removed from the list due to having no online videos, also handled a side case where a video was deleted just after the youtube api reported it as live.

Removed retries in twitch client when the api succesfully returns no live streams. This may need to be reverted if the api is still flakey as it was when that change for retries was first implemented.